server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyjwt==2.10.1
anthropic==0.45.0
gitpython==3.1.43
python-dotenv==1.0.1
